Replace unsafe (int) long casts with Math.toIntExact()
Replace explicit (int) casts of long values with Math.toIntExact() to
detect integer overflow at runtime rather than silently truncating.

Changes cover:
- hemf emf records: HemfComment, HemfFill, HemfPalette, HemfDraw, HemfFont
- hpsf: Section, Array, Vector, Property, PropertySet, PropertySetFactory, VariantSupport
- poifs/crypt: CryptoAPIDecryptor, ChunkedCipherOutputStream, XORDecryptor, XOREncryptor
- poi-ooxml XSSF: XSSFConditionalFormattingRule, XSSFDxfStyleProvider, XSSFOptimiser,
  XSSFChildAnchor, XSSFFontFormatting, XSSFDataBarFormatting, XSSFColor, XSSFTableStyle,
  XSSFWorkbook, XSSFSimpleShape, XSSFName, ColumnHelper
- poi-ooxml XSLF: XSLFSimpleShape, XSLFSheet, XSLFTabStop, XSLFDrawing
- poi-ooxml XDDF: XDDFDataSourcesFactory
- hwpf: PlfLfo, RecordUtil
- util: IOUtils
- poifs/nio: ByteArrayBackedDataSource

Guarded casts (already behind Integer.MAX_VALUE checks) and casts of
floating-point or char values are intentionally left unchanged.

+++ poi-ooxml/src/main/java/org/apache/poi/xslf/usermodel/XSLFDrawing.java
@@ -45,7 +45,7 @@ public class XSLFDrawing {
             // ignore them for now
             if (o instanceof CTNonVisualDrawingProps) {
                 CTNonVisualDrawingProps p = (CTNonVisualDrawingProps)o;
-                sheet.registerShapeId((int)p.getId());
+                sheet.registerShapeId(Math.toIntExact(p.getId()));
             }
         }
     }
